+++ packages/core/src/change_detection/scheduling/zoneless_scheduling_impl.ts
@@ -8,10 +8,10 @@
 
 import {Subscription} from 'rxjs';
 
+import {ɵɵdefineInjectable} from '../../di/interface/defs';
 import {ApplicationRef, ApplicationRefDirtyFlags} from '../../application/application_ref';
-import {Injectable} from '../../di/injectable';
 import {inject} from '../../di/injector_compatibility';
-import {EnvironmentProviders, Provider} from '../../di/interface/provider';
+import {EnvironmentProviders} from '../../di/interface/provider';
 import {makeEnvironmentProviders} from '../../di/provider_collection';
 import {RuntimeError, RuntimeErrorCode, formatRuntimeError} from '../../errors';
 import {PendingTasksInternal} from '../../pending_tasks_internal';
@@ -23,7 +23,6 @@ import {performanceMarkFeature} from '../../util/performance';
 import {NgZone, NgZonePrivate, NoopNgZone, angularZoneInstanceIdProperty} from '../../zone/ng_zone';
 
 import {
-  ChangeDetectionScheduler,
   NotificationSource,
   PROVIDED_ZONELESS,
   SCHEDULE_IN_ROOT_ZONE,
@@ -55,8 +54,14 @@ function trackMicrotaskNotificationForDebugging() {
   }
 }
 
-@Injectable({providedIn: 'root'})
-export class ChangeDetectionSchedulerImpl implements ChangeDetectionScheduler {
+export class ChangeDetectionScheduler {
+  /** @nocollapse */
+  static ɵprov = /** @pureOrBreakMyCode */ /* @__PURE__ */ ɵɵdefineInjectable({
+    token: ChangeDetectionScheduler,
+    providedIn: 'root',
+    factory: () => new ChangeDetectionScheduler(),
+  });
+
   private readonly applicationErrorHandler = inject(INTERNAL_APPLICATION_ERROR_HANDLER);
   private readonly appRef = inject(ApplicationRef);
   private readonly taskService = inject(PendingTasksInternal);
@@ -354,17 +359,10 @@ export function provideZonelessChangeDetection(): EnvironmentProviders {
   }
 
   return makeEnvironmentProviders([
-    ...provideZonelessChangeDetectionInternal(),
+    {provide: NgZone, useClass: NoopNgZone},
+    {provide: ZONELESS_ENABLED, useValue: true},
     typeof ngDevMode === 'undefined' || ngDevMode
       ? [{provide: PROVIDED_ZONELESS, useValue: true}]
       : [],
   ]);
-}
-
-export function provideZonelessChangeDetectionInternal(): Provider[] {
-  return [
-    {provide: ChangeDetectionScheduler, useExisting: ChangeDetectionSchedulerImpl},
-    {provide: NgZone, useClass: NoopNgZone},
-    {provide: ZONELESS_ENABLED, useValue: true},
-  ];
 }
