+++ packages/core/test/bundling/animations-standalone/bundle.golden_symbols.json
@@ -55,7 +55,6 @@
       "CSP_NONCE",
       "ChainedInjector",
       "ChangeDetectionScheduler",
-      "ChangeDetectionSchedulerImpl",
       "ChangeDetectionStrategy",
       "ComponentFactory",
       "ComponentFactory",
@@ -78,7 +77,7 @@
       "DISABLE_ANIMATIONS_FLAG",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DestroyRef",
       "DomAdapter",
@@ -306,7 +305,6 @@
       "_BrowserDomAdapter",
       "_CACHED_BODY",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -324,7 +322,6 @@
       "_IS_WEBKIT",
       "_InjectionToken",
       "_Injector",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -528,7 +525,6 @@
       "eraseStyles",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "execFinalizer",
       "executeCheckHooks",
@@ -816,7 +812,6 @@
       "projectNodes",
       "provideAnimations",
       "provideZonelessChangeDetection",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "publishSignalConfiguration",
