+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -92,7 +92,6 @@
       "CachedInjectorService",
       "ChainedInjector",
       "ChangeDetectionScheduler",
-      "ChangeDetectionSchedulerImpl",
       "ChangeDetectionStrategy",
       "ComponentFactory",
       "ComponentFactory",
@@ -114,7 +113,7 @@
       "DEHYDRATED_VIEWS",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DeferBlockBehavior",
       "DeferBlockInternalState",
       "DeferBlockState",
@@ -272,7 +271,6 @@
       "_AnonymousSubject",
       "_BehaviorSubject",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -283,7 +281,6 @@
       "_EventEmitter_",
       "_InjectionToken",
       "_Injector",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -449,7 +446,6 @@
       "enterView",
       "epoch",
       "errorContext",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "execFinalizer",
       "executeCheckHooks",
@@ -690,7 +686,6 @@
       "profilerCallbacks",
       "projectNodes",
       "provideZonelessChangeDetection",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "publishSignalConfiguration",
