+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -48,7 +48,6 @@
       "CSP_NONCE",
       "ChainedInjector",
       "ChangeDetectionScheduler",
-      "ChangeDetectionSchedulerImpl",
       "ChangeDetectionStrategy",
       "CheckboxControlValueAccessor",
       "CommonModule",
@@ -70,7 +69,7 @@
       "DISABLED",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DefaultIterableDiffer",
       "DefaultIterableDifferFactory",
@@ -303,7 +302,6 @@
       "_BrowserDomAdapter",
       "_BrowserGetTestability",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -337,7 +335,6 @@
       "_NgForOfContext",
       "_NgModuleFactory",
       "_NgModuleRef",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -580,7 +577,6 @@
       "epoch",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "execFinalizer",
       "executeCheckHooks",
@@ -938,7 +934,6 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "providersResolver",
