+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -48,7 +48,6 @@
       "CSP_NONCE",
       "ChainedInjector",
       "ChangeDetectionScheduler",
-      "ChangeDetectionSchedulerImpl",
       "ChangeDetectionStrategy",
       "ChangeDetectorRef",
       "CheckboxControlValueAccessor",
@@ -71,7 +70,7 @@
       "DISABLED",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DefaultIterableDiffer",
       "DefaultIterableDifferFactory",
@@ -307,7 +306,6 @@
       "_BrowserDomAdapter",
       "_BrowserGetTestability",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -339,7 +337,6 @@
       "_NgForOfContext",
       "_NgModuleFactory",
       "_NgModuleRef",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -579,7 +576,6 @@
       "epoch",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "execFinalizer",
       "executeCheckHooks",
@@ -933,7 +929,6 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "providersResolver",
