+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -40,7 +40,6 @@
       "CSP_NONCE",
       "ChainedInjector",
       "ChangeDetectionScheduler",
-      "ChangeDetectionSchedulerImpl",
       "ChangeDetectionStrategy",
       "ComponentFactory",
       "ComponentFactory",
@@ -59,7 +58,7 @@
       "DISCONNECTED_NODES",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DestroyRef",
       "DomAdapter",
@@ -247,7 +246,6 @@
       "_BehaviorSubject",
       "_BrowserDomAdapter",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -267,7 +265,6 @@
       "_HttpResponseBase",
       "_InjectionToken",
       "_Injector",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -464,7 +461,6 @@
       "epoch",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "escapeCommentText",
       "execFinalizer",
@@ -631,7 +627,7 @@
       "isInI18nBlock",
       "isInInjectionContext",
       "isInSkipHydrationBlock",
-      "isInSkipHydrationBlock2",
+      "isInSkipHydrationBlock",
       "isInlineTemplate",
       "isInputBinding",
       "isInteropObservable",
@@ -730,7 +726,6 @@
       "profilerCallbacks",
       "projectNodes",
       "provideClientHydration",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "publishSignalConfiguration",
