+++ packages/platform-browser/testing/src/browser.ts
@@ -11,7 +11,6 @@ import {
   createPlatformFactory,
   NgModule,
   StaticProvider,
-  ɵprovideZonelessChangeDetectionInternal as provideZonelessChangeDetectionInternal,
   ɵinternalProvideZoneChangeDetection as internalProvideZoneChangeDetection,
   PlatformRef,
 } from '@angular/core';
@@ -38,7 +37,6 @@ export const platformBrowserTesting: (extraProviders?: StaticProvider[]) => Plat
   exports: [BrowserModule],
   providers: [
     {provide: APP_ID, useValue: 'a'},
-    provideZonelessChangeDetectionInternal(),
     ZONELESS_BY_DEFAULT ? [] : internalProvideZoneChangeDetection({}),
     ɵprovideFakePlatformNavigation(),
     {provide: TestComponentRenderer, useClass: DOMTestComponentRenderer},
